% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForestPlot.R
\name{ForestPlot}
\alias{ForestPlot}
\title{Plot the estimated subgroup causal effects via forest plot}
\usage{
ForestPlot(obj, logodds = FALSE, xlab = NULL, legend_label = NULL, ...)
}
\arguments{
\item{obj}{a list of \code{summary.PSweight_sga} objects obtained with \code{\link{summary.PSweight_sga}} function.
This is done to plot multiple PS weighting methods.}

\item{logodds}{an indicator to specify whether the effects are estimated on the logodds scale. The default is FALSE.}

\item{xlab}{an optional character string indicating the xlab of the plot.}

\item{legend_label}{an optional character vector indicating the PS weighting methods corresponding to the \code{summary.PSweight_sga} object.
The default is \code{method} returned from the \code{summary.PSweight_sga} function.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Plot of the indicated type.
}
\description{
Plot the estimated subgroup causal effects via forest plot
}
